規模の情報をIntではなくDoubleで保持するように変更

+++ src/main/scala/pdashdata/AdditionalParts.scala
@@ -2,8 +2,8 @@ package pdashdata
 
 class AdditionalParts(
     val name: String,
-    val estSize: Int,
-    val actSize: Int,
+    val estSize: Double,
+    val actSize: Double,
     val relativeSize: String,
     val estNumOfMethod: Int,
     val estIsNewReused: Boolean,
