+++ src/main/scala/pdashdata/ReusedParts.scala
@@ -2,7 +2,7 @@ package pdashdata
 
 class ReusedParts(
     val name: String,
-    val estSize: Int,
-    val actSize: Int,
+    val estSize: Double,
+    val actSize: Double,
 ) {
 }
